simplify async mocking for 3.5-3.7
