+++ sdk/identity/azure-identity/tests/helpers.py
@@ -141,50 +141,3 @@ def urlsafeb64_decode(s):
 
     padding_needed = 4 - len(s) % 4
     return base64.urlsafe_b64decode(s + b"=" * padding_needed)
-
-
-try:
-    import asyncio
-    import functools
-
-    def get_completed_future(result=None):
-        future = asyncio.Future()
-        future.set_result(result)
-        return future
-
-    def wrap_in_future(fn):
-        """Return a completed Future whose result is the return of fn.
-
-        Added to simplify using unittest.Mock in async code. Python 3.8's AsyncMock would be preferable.
-        """
-
-        @functools.wraps(fn)
-        def wrapper(*args, **kwargs):
-            result = fn(*args, **kwargs)
-            return get_completed_future(result)
-        return wrapper
-
-
-    class AsyncMockTransport(mock.MagicMock):
-        """MagicMock with do-nothing aenter/exit, for mocking async transport.
-
-        aenter/exit here return completed Futures so they needn't be declared async. If they were async functions,
-        2.7 would raise SyntaxError on importing this module.
-
-        3.9.0, or a future version of 3.8, will obviate this class by implementing aenter/exit on MagicMock
-        (https://bugs.python.org/issue38093).
-        """
-
-        def __aenter__(self):
-            return get_completed_future()
-
-        def __aexit__(self, *exc_info):
-            return get_completed_future()
-
-    def async_validating_transport(requests, responses):
-        sync_transport = validating_transport(requests, responses)
-        return AsyncMockTransport(send=wrap_in_future(sync_transport.send))
-
-
-except ImportError:
-    pass
