+++ sdk/identity/azure-identity/tests/test_msi_credential_async.py
@@ -8,7 +8,7 @@ from azure.identity._constants import EnvironmentVariables
 from azure.identity.aio._credentials.managed_identity import MsiCredential
 import pytest
 
-from helpers import AsyncMockTransport
+from helpers_async import AsyncMockTransport
 
 
 @pytest.mark.asyncio
@@ -20,7 +20,7 @@ async def test_close():
 
     await credential.close()
 
-    assert transport.__aexit__.call_count == 1
+    assert transport.exited
 
 
 @pytest.mark.asyncio
@@ -33,4 +33,4 @@ async def test_context_manager():
     async with credential:
         pass
 
-    assert transport.__aexit__.call_count == 1
+    assert transport.exited
