+++ sdk/identity/azure-identity/tests/test_shared_cache_credential_async.py
@@ -19,15 +19,8 @@ from azure.identity._internal.user_agent import USER_AGENT
 from msal import TokenCache
 import pytest
 
-from helpers import (
-    AsyncMockTransport,
-    async_validating_transport,
-    build_aad_response,
-    build_id_token,
-    mock_response,
-    Request,
-)
-
+from helpers import build_aad_response, build_id_token, mock_response, Request
+from helpers_async import async_validating_transport, AsyncMockTransport
 from test_shared_cache_credential import get_account_event, populated_cache
 
 
@@ -38,7 +31,7 @@ async def test_close():
 
     await credential.close()
 
-    assert transport.__aexit__.call_count == 1
+    assert transport.exited
 
 
 @pytest.mark.asyncio
@@ -49,7 +42,7 @@ async def test_context_manager():
     async with credential:
         pass
 
-    assert transport.__aexit__.call_count == 1
+    assert transport.exited
 
 
 @pytest.mark.asyncio
